Write and implement JournalPolicy rules

+++ app/Http/Controllers/JournalController.php
@@ -34,7 +34,12 @@ class JournalController extends Controller
      */
     public function create()
     {
-        return view('journal.create');
+        if (Auth::user()->can('create', Journal::class)) {
+            return view('journal.create');
+        }
+
+        // Redirect to journal index
+        return redirect()->route('journal.index');
     }
 
     /**
@@ -45,19 +50,24 @@ class JournalController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'title' => 'required|max:255'
-        ]);
+        if (Auth::user()->can('create', Journal::class)) {
+            $request->validate([
+                'title' => 'required|max:255'
+            ]);
 
-        $journal = new Journal;
-        $journal->title = $request->title;
-        $journal->description = $request->description;
-        $journal->creator_id = $journal->current_user_id = Auth::id();
-        $journal->save();
-        $journal->users()->save(Auth::user());
+            $journal = new Journal;
+            $journal->title = $request->title;
+            $journal->description = $request->description;
+            $journal->creator_id = $journal->current_user_id = Auth::id();
+            $journal->save();
+            $journal->users()->save(Auth::user());
 
-        $request->session()->flash('status', "<strong>{$journal->title}</strong> has been created.");
-        return redirect()->route('journal.show', compact('journal'));
+            $request->session()->flash('status', "<strong>{$journal->title}</strong> has been created.");
+            return redirect()->route('journal.show', compact('journal'));
+        }
+
+        // Redirect to journal index
+        return redirect()->route('journal.index');
     }
 
     /**
@@ -70,7 +80,7 @@ class JournalController extends Controller
     public function show(Request $request, Journal $journal)
     {
         // Show the journal to the current user only.
-        if (Auth::id() == $journal->current_user->id) {
+        if (Auth::user()->can('view', $journal)) {
             $drafts = $journal->entries()->where('status', 'draft')->paginate(5);
             $drafts->withPath(route('journal.show', $journal));
 
@@ -96,7 +106,7 @@ class JournalController extends Controller
     public function edit(Request $request, Journal $journal)
     {
         // Only the journal creator can edit journal details
-        if (Auth::id() == $journal->creator->id) {
+        if (Auth::user()->can('edit', $journal)) {
             return view('journal.edit', compact('journal'));
         }
 
@@ -118,16 +128,21 @@ class JournalController extends Controller
      */
     public function update(Request $request, Journal $journal)
     {
-        $request->validate([
-            'title' => 'required|max:255'
-        ]);
+        if (Auth::user()->can('edit', $journal)) {
+            $request->validate([
+                'title' => 'required|max:255'
+            ]);
 
-        $journal->title = $request->title;
-        $journal->description = $request->description;
-        $journal->save();
+            $journal->title = $request->title;
+            $journal->description = $request->description;
+            $journal->save();
 
-        $request->session()->flash('status', "<strong>{$journal->title}</strong> has been updated.");
-        return redirect()->route('journal.show', compact('journal'));
+            $request->session()->flash('status', "<strong>{$journal->title}</strong> has been updated.");
+            return redirect()->route('journal.show', compact('journal'));
+        }
+
+        // Redirect to journal index
+        return redirect()->route('journal.index');
     }
 
     /**
@@ -139,17 +154,20 @@ class JournalController extends Controller
      */
     public function destroy(Request $request, Journal $journal)
     {
-        // TODO: Validate request
+        if (Auth::user()->can('delete', $journal)) {
 
-        $journal->delete();
-        $request->session()->flash('status',
-            "<strong>{$journal->title}</strong> has been deleted.
-            <a href=\"#\" onclick=\"event.preventDefault(); document.getElementById('undo-form').submit();\">Undo</a>
-            <form id=\"undo-form\" style=\"display:none;\" method=\"post\" action='/journal/undodelete'>
-                <input type='hidden' name='_token' value='" . csrf_token() . "'>
-                <input type='hidden' name='journal_id' value='{$journal->id}'>
-            </form>"
-        );
+            $journal->delete();
+            $request->session()->flash('status',
+                "<strong>{$journal->title}</strong> has been deleted.
+                <a href=\"#\" onclick=\"event.preventDefault(); document.getElementById('undo-form').submit();\">Undo</a>
+                <form id=\"undo-form\" style=\"display:none;\" method=\"post\" action='/journal/undodelete'>
+                    <input type='hidden' name='_token' value='" . csrf_token() . "'>
+                    <input type='hidden' name='journal_id' value='{$journal->id}'>
+                </form>"
+            );
+        }
+
+        // Redirect to journal index
         return redirect()->route('journal.index');
     }
 
@@ -161,9 +179,13 @@ class JournalController extends Controller
      */
     public function confirmDelete(Journal $journal)
     {
-        // TODO: Validate request
+        if (Auth::user()->can('delete', $journal)) {
 
-        return view('journal.confirmDelete', compact('journal'));
+            return view('journal.confirmDelete', compact('journal'));
+        }
+
+        // Redirect to journal index
+        return redirect()->route('journal.index');
     }
 
     /**
@@ -174,11 +196,14 @@ class JournalController extends Controller
      */
     public function undoDelete(Request $request)
     {
-        // TODO: Validate request
+        if (Auth::user()->can('restore', $journal)) {
 
-        $journal = \App\Journal::withTrashed()->find($request->input('journal_id'));
-        $journal->restore();
-        $request->session()->flash('status', "<strong>{$journal->title}</strong> has been restored.");
+            $journal = \App\Journal::withTrashed()->find($request->input('journal_id'));
+            $journal->restore();
+            $request->session()->flash('status', "<strong>{$journal->title}</strong> has been restored.");
+        }
+
+        // Redirect to journal index
         return redirect()->route('journal.index');
     }
 
@@ -190,9 +215,12 @@ class JournalController extends Controller
      */
     public function invite(Journal $journal)
     {
-        // TODO: Validate request
+        if (Auth::user()->can('invite', $journal)) {
+            return view('journal.invite', compact('journal'));
+        }
 
-        return view('journal.invite', compact('journal'));
+        // Redirect to journal index
+        return redirect()->route('journal.index');
     }
 
     /**
@@ -203,11 +231,20 @@ class JournalController extends Controller
      */
     public function contents(Journal $journal)
     {
-        // TODO: Validate request
-        $entries = $journal->entries()->where('status', 'final')->paginate(2);
-        $entries->withPath(route('journal.contents', $journal));
+        if (Auth::user()->can('view', $journal)) {
+            $entries = $journal->entries()->where('status', 'final')->paginate(2);
+            $entries->withPath(route('journal.contents', $journal));
 
-        return view('entry.index', compact('journal', 'entries'));
+            return view('entry.index', compact('journal', 'entries'));
+        }
+
+        // Show a flash message if the user belongs to the journal.
+        if (Auth::user()->isInJournal($journal)) {
+            $request->session()->flash('status', "{$journal->current_user->name} has <strong>{$journal->title}</strong> right now. You'll be able to view it when it's your turn.");
+        }
+
+        // Redirect to journal index
+        return redirect()->route('journal.index');
     }
 
     /**
@@ -217,6 +254,16 @@ class JournalController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function addEntry(Journal $journal) {
-        return view('entry.create', compact('journal'));
+        if (Auth::user()->can('addEntry', $journal)) {
+            return view('entry.create', compact('journal'));
+        }
+
+        // Show a flash message if the user belongs to the journal.
+        if (Auth::user()->isInJournal($journal)) {
+            $request->session()->flash('status', "{$journal->current_user->name} has <strong>{$journal->title}</strong> right now. You'll be able to write in it when it's your turn.");
+        }
+
+        // Redirect to journal index
+        return redirect()->route('journal.index');
     }
 }
